feat: configure Routing
Implements [React Router Dom] to configure Routing. `react-router-dom`
is used to build single-page applications (applications that have many
pages or components but the page is never refreshed instead the content
is dynamically fetched based on the URL).

The major advantage of `react-router-dome` is that the page doesn't have
to be refreshed when a link to another page is clicked. It's very
fast compared to traditional page navigation.

Wrap up `Navbar` and `Home` in `Router`.

Issue retrotechie/rt-hris#31
Link:
- [React Router Dom](https://www.npmjs.com/package/react-router-dom)
Reference:
- [What is React Router Dome](https://www.geeksforgeeks.org/what-is-react-router-dom/)