feat: store `Employee` info in `state`
`useState` hook is used to create states that can be updated and
rendered in UI. It returns an array with 2 elements:
- 1st element: the current state value (`employee`)
- 2nd element: function to update that state (`setEmployee`), this will
trigger a re-render of the component.

[Object destructuring] is a way to extract properties from an obj and
assign them to separate variables. The properties `firstName`,
`lastName` and `email` from the `employee` obj are being extracted and
assigned to separate variables with the same name. This allows the
component to reference these properties directly, rather than having to
use `employee.firstName`, `employee.lastName`, and `employee.email`.

[Spread operator] is used to update the `employee` state. It creates a new
obj using the spread operator to spread the properties of the current
`employee` obj. Then, it adds a new property with a computed property
name `[e.target.name]` and the value of `e.target.value`. Finally, it
updates the state with the new object. This allows the componentto
update only the property that has changed, while preserving the values
of the other properties.

Issue retrotechie/rt-hris#39
Links:
- [Object destructuring](https://www.javascripttutorial.net/es6/javascript-object-destructuring/)
- [Spread operator](https://www.w3schools.com/react/react_es6_spread.asp)
Reference:
- [State Hook](https://reactjs.org/docs/hooks-state.html)